feat(ui): reuse DB selector between Users and Roles pages
